move sasCodeHelper
